Fixed Start Button - Began Server Browser
Start button will now load the player into the game when there is no xml data. The loaded characters have no customization. The ball is throwing weirdly. With flicky wrist movements it rockets away in a direction not relative to the movement. Also added a slider to the level select value in the inspector.