Add fade to black and return to home()
Added a UI element that sits in front of the camera. The UI element is a child of the main camera. It is controlled by the script FadeToBlack.cs. The functions in fade to black are called in GameManger.cs and ServerBrowser.cs. I also added an IEnumerator to the GameManager that should return the player to the house, but I have not tested it yet. The IEnumerator is called from AI.cs in the else of the ThrowBall() function.